division

+++ programming_paradigm/robust_division_calculator.py
@@ -1,7 +1,7 @@
 def safe_divide(numerator, denominator) :
     try:
-        num1 = int(numerator)
-        num2 = int(denominator)
+        num1 = float(numerator)
+        num2 = float(denominator)
         result = num1 / num2
         return f"The result of the division is {result}"
     except ZeroDivisionError:
